requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
httpx==0.28.1
redis==5.2.1
asyncpg==0.30.0
pydantic-settings==2.7.1
networkx==3.4.2
google-genai==1.3.0
python-json-logger==3.2.1

